requirements.txt
Flask==3.0.3
Werkzeug==3.0.2
Jinja2==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
flask-smorest==0.44.0
marshmallow==3.21.2
flask-smorest==0.44.0
marshmallow==3.21.2
pymongo
flask-jwt-extended
python-dotenv
bcrypt
marshmallow_enum
pytest
pytest-flask